Changed code to use averaged sensor data. For the time being no data gets set but only because the required function never gets called in main.

+++ sensor_station/include/models/sensor_data.h
@@ -9,8 +9,8 @@ typedef struct {
 		uint16_t air_humidity;
 		uint32_t air_pressure;
 		uint16_t air_quality;
-		uint16_t temperature;
-		uint8_t light_intensity;
+		uint8_t temperature;
+		uint16_t light_intensity;
 } sensor_data_t;
 
 #endif
